pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "march-madness-moe"
version = "0.1.0"
description = "March Madness prediction system using Mixture of Experts architecture"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "xgboost",
    "torch",
    "kenpompy",
    "torchsort",
    "jupyterlab",
    "optuna",
    "shap",
    "tqdm",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
